requirements.txt
manga-ocr
deep-translator
Pillow
customtkinter
transformers
llama-cpp-python
huggingface-hub